Refactor method/class name to reflect meaning
